Qiskit to perceval converter is now a class
